feat(KAN-192): maj architecture package
