+++ src/main/java/com/wcs/travel_blog/article/repository/ArticleRepository.java
@@ -1,6 +1,6 @@
-package com.wcs.travel_blog.repository;
+package com.wcs.travel_blog.article.repository;
 
-import com.wcs.travel_blog.model.Article;
+import com.wcs.travel_blog.article.model.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArticleRepository extends JpaRepository<Article,Long> {
