+++ src/main/java/com/wcs/travel_blog/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
-package com.wcs.travel_blog.repository;
+package com.wcs.travel_blog.comment.repository;
 
-import com.wcs.travel_blog.model.Comment;
+import com.wcs.travel_blog.comment.model.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
