+++ src/main/java/com/wcs/travel_blog/media/repository/MediaRepository.java
@@ -1,6 +1,6 @@
-package com.wcs.travel_blog.repository;
+package com.wcs.travel_blog.media.repository;
 
-import com.wcs.travel_blog.model.Media;
+import com.wcs.travel_blog.media.model.Media;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MediaRepository extends JpaRepository<Media, Long> {
