+++ src/main/java/com/wcs/travel_blog/step/repository/StepRepository.java
@@ -1,6 +1,6 @@
-package com.wcs.travel_blog.repository;
+package com.wcs.travel_blog.step.repository;
 
-import com.wcs.travel_blog.model.Step;
+import com.wcs.travel_blog.step.model.Step;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StepRepository extends JpaRepository<Step, Long> {
