+++ src/main/java/com/wcs/travel_blog/theme/model/Theme.java
@@ -1,5 +1,6 @@
-package com.wcs.travel_blog.model;
+package com.wcs.travel_blog.theme.model;
 
+import com.wcs.travel_blog.step.model.Step;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
