+++ src/main/java/com/wcs/travel_blog/theme/repository/ThemeRepository.java
@@ -1,6 +1,6 @@
-package com.wcs.travel_blog.repository;
+package com.wcs.travel_blog.theme.repository;
 
-import com.wcs.travel_blog.model.Theme;
+import com.wcs.travel_blog.theme.model.Theme;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ThemeRepository extends JpaRepository<Theme, Long> {
