+++ src/main/java/com/wcs/travel_blog/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.wcs.travel_blog.repository;
+package com.wcs.travel_blog.user.repository;
 
-import com.wcs.travel_blog.model.User;
+import com.wcs.travel_blog.user.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
